+++ cli/config.go
@@ -18,12 +18,11 @@ import (
 )
 
 var (
-	traceStorage       = app.Flag("trace-storage", "Enables tracing of storage operations.").Default("true").Hidden().Bool()
-	traceObjectManager = app.Flag("trace-object-manager", "Enables tracing of object manager operations.").Envar("KOPIA_TRACE_OBJECT_MANAGER").Bool()
-	traceLocalFS       = app.Flag("trace-localfs", "Enables tracing of local filesystem operations").Envar("KOPIA_TRACE_FS").Bool()
-	enableCaching      = app.Flag("caching", "Enables caching of objects (disable with --no-caching)").Default("true").Hidden().Bool()
-	enableListCaching  = app.Flag("list-caching", "Enables caching of list results (disable with --no-list-caching)").Default("true").Hidden().Bool()
-	metricsListenAddr  = app.Flag("metrics-listen-addr", "Expose Prometheus metrics on a given host:port").Hidden().String()
+	traceStorage      = app.Flag("trace-storage", "Enables tracing of storage operations.").Default("true").Hidden().Bool()
+	traceLocalFS      = app.Flag("trace-localfs", "Enables tracing of local filesystem operations").Envar("KOPIA_TRACE_FS").Bool()
+	enableCaching     = app.Flag("caching", "Enables caching of objects (disable with --no-caching)").Default("true").Hidden().Bool()
+	enableListCaching = app.Flag("list-caching", "Enables caching of list results (disable with --no-list-caching)").Default("true").Hidden().Bool()
+	metricsListenAddr = app.Flag("metrics-listen-addr", "Expose Prometheus metrics on a given host:port").Hidden().String()
 
 	configPath = app.Flag("config-file", "Specify the config file to use.").Default(defaultConfigFileName()).Envar("KOPIA_CONFIG_PATH").String()
 )
@@ -77,10 +76,6 @@ func applyOptionsFromFlags(ctx context.Context, opts *repo.Options) *repo.Option
 
 	if *traceStorage {
 		opts.TraceStorage = log(ctx).Debugf
-	}
-
-	if *traceObjectManager {
-		opts.ObjectManagerOptions.Trace = log(ctx).Debugf
 	}
 
 	return opts
