+++ repo/object/object_reader.go
@@ -1,19 +1,37 @@
 package object
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
 	"io"
 
 	"github.com/pkg/errors"
+
+	"github.com/kopia/kopia/repo/compression"
+	"github.com/kopia/kopia/repo/content"
 )
 
-func (i *indirectObjectEntry) endOffset() int64 {
-	return i.Start + i.Length
+// Open creates new ObjectReader for reading given object from a repository.
+func Open(ctx context.Context, r contentReader, objectID ID) (Reader, error) {
+	return openAndAssertLength(ctx, r, objectID, -1)
+}
+
+// VerifyObject ensures that all objects backing ObjectID are present in the repository
+// and returns the content IDs of which it is composed.
+func VerifyObject(ctx context.Context, cr contentReader, oid ID) ([]content.ID, error) {
+	tracker := &contentIDTracker{}
+
+	if err := verifyObjectInternal(ctx, cr, oid, tracker); err != nil {
+		return nil, err
+	}
+
+	return tracker.contentIDs(), nil
 }
 
 type objectReader struct {
-	ctx  context.Context
-	repo *Manager
+	ctx context.Context
+	cr  contentReader
 
 	seekTable []indirectObjectEntry
 
@@ -79,7 +97,7 @@ func (r *objectReader) Read(buffer []byte) (int, error) {
 func (r *objectReader) openCurrentChunk() error {
 	st := r.seekTable[r.currentChunkIndex]
 
-	rd, err := r.repo.openAndAssertLength(r.ctx, st.Object, st.Length)
+	rd, err := openAndAssertLength(r.ctx, r.cr, st.Object, st.Length)
 	if err != nil {
 		return err
 	}
@@ -171,4 +189,150 @@ func (r *objectReader) Close() error {
 
 func (r *objectReader) Length() int64 {
 	return r.totalLength
+}
+
+func openAndAssertLength(ctx context.Context, cr contentReader, objectID ID, assertLength int64) (Reader, error) {
+	if indexObjectID, ok := objectID.IndexObjectID(); ok {
+		// recursively calls openAndAssertLength
+		seekTable, err := loadSeekTable(ctx, cr, indexObjectID)
+		if err != nil {
+			return nil, err
+		}
+
+		totalLength := seekTable[len(seekTable)-1].endOffset()
+
+		return &objectReader{
+			ctx:         ctx,
+			cr:          cr,
+			seekTable:   seekTable,
+			totalLength: totalLength,
+		}, nil
+	}
+
+	return newRawReader(ctx, cr, objectID, assertLength)
+}
+
+func verifyIndirectObjectInternal(ctx context.Context, cr contentReader, indexObjectID ID, tracker *contentIDTracker) error {
+	if err := verifyObjectInternal(ctx, cr, indexObjectID, tracker); err != nil {
+		return errors.Wrap(err, "unable to read index")
+	}
+
+	seekTable, err := loadSeekTable(ctx, cr, indexObjectID)
+	if err != nil {
+		return err
+	}
+
+	for _, m := range seekTable {
+		err := verifyObjectInternal(ctx, cr, m.Object, tracker)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func verifyObjectInternal(ctx context.Context, r contentReader, oid ID, tracker *contentIDTracker) error {
+	if indexObjectID, ok := oid.IndexObjectID(); ok {
+		return verifyIndirectObjectInternal(ctx, r, indexObjectID, tracker)
+	}
+
+	if contentID, _, ok := oid.ContentID(); ok {
+		if _, err := r.ContentInfo(ctx, contentID); err != nil {
+			return errors.Wrapf(err, "error getting content info for %v", contentID)
+		}
+
+		tracker.addContentID(contentID)
+
+		return nil
+	}
+
+	return errors.Errorf("unrecognized object type: %v", oid)
+}
+
+type indirectObject struct {
+	StreamID string                `json:"stream"`
+	Entries  []indirectObjectEntry `json:"entries"`
+}
+
+func loadSeekTable(ctx context.Context, cr contentReader, indexObjectID ID) ([]indirectObjectEntry, error) {
+	r, err := openAndAssertLength(ctx, cr, indexObjectID, -1)
+	if err != nil {
+		return nil, err
+	}
+	defer r.Close() //nolint:errcheck
+
+	var ind indirectObject
+
+	if err := json.NewDecoder(r).Decode(&ind); err != nil {
+		return nil, errors.Wrap(err, "invalid indirect object")
+	}
+
+	return ind.Entries, nil
+}
+
+func newRawReader(ctx context.Context, cr contentReader, objectID ID, assertLength int64) (Reader, error) {
+	contentID, compressed, ok := objectID.ContentID()
+	if !ok {
+		return nil, errors.Errorf("unsupported object ID: %v", objectID)
+	}
+
+	payload, err := cr.GetContent(ctx, contentID)
+	if errors.Is(err, content.ErrContentNotFound) {
+		return nil, errors.Wrapf(ErrObjectNotFound, "content %v not found", contentID)
+	}
+
+	if err != nil {
+		return nil, errors.Wrap(err, "unexpected content error")
+	}
+
+	if compressed {
+		var b bytes.Buffer
+
+		if err = decompress(&b, payload); err != nil {
+			return nil, errors.Wrap(err, "decompression error")
+		}
+
+		payload = b.Bytes()
+	}
+
+	if assertLength != -1 && int64(len(payload)) != assertLength {
+		return nil, errors.Errorf("unexpected chunk length %v, expected %v", len(payload), assertLength)
+	}
+
+	return newObjectReaderWithData(payload), nil
+}
+
+func decompress(output *bytes.Buffer, b []byte) error {
+	compressorID, err := compression.IDFromHeader(b)
+	if err != nil {
+		return errors.Wrap(err, "invalid compression header")
+	}
+
+	compressor := compression.ByHeaderID[compressorID]
+	if compressor == nil {
+		return errors.Errorf("unsupported compressor %x", compressorID)
+	}
+
+	return compressor.Decompress(output, b)
+}
+
+type readerWithData struct {
+	io.ReadSeeker
+	length int64
+}
+
+func (rwd *readerWithData) Close() error {
+	return nil
+}
+
+func (rwd *readerWithData) Length() int64 {
+	return rwd.length
+}
+
+func newObjectReaderWithData(data []byte) Reader {
+	return &readerWithData{
+		ReadSeeker: bytes.NewReader(data),
+		length:     int64(len(data)),
+	}
 }
